Make distane_to_primary_level a funtion, not a method
